Module08 TP

+++ src/Controller/WishController.php
@@ -11,11 +11,11 @@ use Symfony\Component\Routing\Attribute\Route;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
-use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Component\String\Slugger\SluggerInterface;
 use App\Entity\Category;
-
+use Dom\Comment;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/wish', name: 'wish_')]
 final class WishController extends AbstractController
@@ -87,6 +87,7 @@ final class WishController extends AbstractController
     }
 
     #[Route('/{id}/modify', name: "modify_form", methods: ['GET', 'POST'])]
+    #[IsGranted('ROLE_USER')]
     public function update(
         Wish $wish,
         Request $request,
@@ -94,49 +95,57 @@ final class WishController extends AbstractController
         SluggerInterface $slugger,
         #[Autowire('%kernel.project_dir%/public/uploads/img')] string $imageDirectory
     ): Response {
-        $editForm = $this->createForm(WishType::class, $wish);
-        $editForm->handleRequest($request);
+        if ($this->getUser() === $wish->getUser()) {
+            $editForm = $this->createForm(WishType::class, $wish);
+            $editForm->handleRequest($request);
 
-        if ($editForm->isSubmitted() && $editForm->isValid()) {
-            // Récupération du fichier après handleRequest
-            $image = $editForm->get('image')->getData();
+            if ($editForm->isSubmitted() && $editForm->isValid()) {
+                $image = $editForm->get('image')->getData();
 
-            if ($image) {
-                $originalFilename = pathinfo($image->getClientOriginalName(), PATHINFO_FILENAME);
-                $safeFilename = $slugger->slug($originalFilename);
-                $newFilename = $safeFilename . '-' . uniqid() . '.' . $image->guessExtension();
+                if ($image) {
+                    $originalFilename = pathinfo($image->getClientOriginalName(), PATHINFO_FILENAME);
+                    $safeFilename = $slugger->slug($originalFilename);
+                    $newFilename = $safeFilename . '-' . uniqid() . '.' . $image->guessExtension();
 
-                try {
-                    $image->move($imageDirectory, $newFilename);
-                } catch (FileException $e) {
-                    $this->addFlash('error', 'File upload failed: ' . $e->getMessage());
+                    try {
+                        $image->move($imageDirectory, $newFilename);
+                    } catch (FileException $e) {
+                        $this->addFlash('error', 'File upload failed: ' . $e->getMessage());
+                    }
+
+                    $wish->setImageName($newFilename);
                 }
 
-                $wish->setImageName($newFilename);
+                $em->persist($wish);
+                $em->flush();
+
+                $this->addFlash('success', 'Idea successfully updated!');
+                return $this->redirectToRoute('wish_show', ['id' => $wish->getId()]);
             }
 
-            $em->persist($wish);
-            $em->flush();
-
-            $this->addFlash('success', 'Idea successfully updated!');
-            return $this->redirectToRoute('wish_show', ['id' => $wish->getId()]);
+            return $this->render('wish/edit-form.html.twig', [
+                'editForm' => $editForm,
+                'wish' => $wish,
+            ]);
         }
 
-        return $this->render('wish/edit-form.html.twig', [
-            'editForm' => $editForm,
-            'wish' => $wish
-
-        ]);
+        $this->addFlash('error', 'You can\'t modify this wish !');
+        return $this->redirectToRoute('wish_show', ['id' => $wish->getId()]);
     }
 
+    #[IsGranted('ROLE_USER', 'ROLE_ADMIN')]
     #[Route('/{id}/delete', name: 'delete', requirements: ['id' => '\d+'], methods: ['GET'])]
     public function delete(int $id, WishRepository $wishRepository, EntityManagerInterface $em): Response
     {
         $wish = $wishRepository->find($id);
-
         if (!$wish) {
             $this->addFlash('error', 'Wish not found');
             return $this->redirectToRoute('wish_list');
+        }
+
+        if ($this->getUser() !== $wish->getUser()) {
+            $this->addFlash('error', 'You can\'t delete this wish !');
+            return $this->redirectToRoute('wish_show', ['id' => $wish->getId()]);
         }
 
         $em->remove($wish);
@@ -147,4 +156,31 @@ final class WishController extends AbstractController
         return $this->redirectToRoute('wish_list');
     }
 
+    #[IsGranted('ROLE_USER')]
+    #[Route('/{id}/add-comment', name: 'comment', requirements: ['id' => '\d+'], methods: ['GET', 'POST'])]
+    public function comment(int $id, Comment $comment, WishRepository $wishRepository, EntityManagerInterface $em): Response
+    {
+        $wish = $wishRepository->find($id);
+
+        if (!$wish) {
+            $this->addFlash('error', 'Wish not found');
+            return $this->redirectToRoute('wish_list');
+        }
+
+        if ($this->getUser() !== $wish->getUser()) {
+
+            $wish->setComment($comment);
+
+            return $this->render('wish/detail.html.twig', [
+                'comment' => $comment
+            ]);
+        }
+
+        $em->remove($wish);
+        $em->flush();
+
+        $this->addFlash('success', 'Idea successfully removed');
+
+        return $this->redirectToRoute('wish_list');
+    }
 }
